+++ lib/blocs/subbudget_bloc/subbudget_bloc.dart
@@ -70,7 +70,6 @@ class SubbudgetBloc extends Bloc<SubbudgetEvent, SubbudgetState> {
       });
     });
 
-    // TODO hier weitermachen und Subbudgets erstellen / bearbeiten implementieren extra Bloc für Subbudgets erstellen?!
     on<LoadSubbudgetEvent>((event, emit) async {
       emit(SubbudgetLoadingState());
       CategorieInputFieldCubit categorie = BlocProvider.of<CategorieInputFieldCubit>(event.context);
